started building out the navigation bar
