+++ pages/_app.tsx
@@ -1,7 +1,7 @@
 import 'tailwindcss/tailwind.css';
 import { AppProps } from 'next/app';
 import Head from 'next/head';
-import { NavigationBar } from '../components/NavigationBar';
+import { NavigationBar } from '../components/NavigationBar/NavigationBar';
 
 function App({ Component, pageProps }: AppProps) {
     return (
